Refactor accordion components; update imports and remove unused sections in templates.
