+++ src/app/Shared/components/accordionControl/accordion-section/accordion-section.component.scss
