+++ src/app/Shared/components/accordionControl/saved-group-accordion/saved-group-accordion.component.scss
