+++ src/app/Shared/components/testing-level/testing-level.component.html
@@ -2,9 +2,7 @@
 </app-table-dropdown> -->
 <br>
 <br>
-<app-accordion-section *ngFor="let section of sections" [section]="section" [parentPath]="''"
-  (fieldSelected)="onFieldSelected($event)">
-</app-accordion-section>
+
 <br>
 <br>
 <!-- Use the QueryTable component here -->
